[Chore] Merge duplicated api folders into one
