+++ src/api/user.api.ts
@@ -1,4 +1,4 @@
-import axiosInstance from '@/app/api/axiosInstance';
+import axiosInstance from '@/api/axiosInstance';
 
 export const getCurrentUser = async () => {
   const response = await axiosInstance.get('/user');
