+++ src/app/(pages)/(auth)/signup/signupForm.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import { signup } from '../../../api/auth';
+import { signup } from '../../../../api/auth';
 import { useRouter } from 'next/navigation';
 import { toast } from 'react-toastify';
 import { AxiosError } from 'axios';
