+++ src/hooks/useCurrentUser.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getCurrentUser } from '@/app/api/user.api';
+import { getCurrentUser } from '@/api/user.api';
 import { UserResponse } from '@/types/usertypes';
 
 export const useCurrentUser = () => {
